11: Implement a transport protocol-agnostic framework for gateway servers  * Change type of all milliseconds to long. Status: Incomplete

+++ gateway/netty-gateway-support/src/main/java/vnet/sms/gateway/nettysupport/login/incoming/IncomingLoginRequestsChannelHandler.java
@@ -40,13 +40,13 @@ public class IncomingLoginRequestsChannelHandler<ID extends Serializable>
 
 	private final AtomicReference<Authentication>	authenticatedClient	   = new AtomicReference<Authentication>();
 
-	private final int	                          failedLoginResponseDelayMillis;
+	private final long	                          failedLoginResponseDelayMillis;
 
 	private final Timer	                          failedLoginResponseTimer	= new HashedWheelTimer();
 
 	public IncomingLoginRequestsChannelHandler(
 	        final AuthenticationManager authenticationManager,
-	        final int failedLoginResponseDelayMillis) {
+	        final long failedLoginResponseDelayMillis) {
 		notNull(authenticationManager,
 		        "Argument 'authenticationManager' must not be null");
 		this.authenticationManager = authenticationManager;
